Import UIKit in ChameleonStatusBar.h and NSArray+Chameleon.h
I tried integrating Chameleon with my project running on iOS 8 with the xCode 6 beta.
I was using Swift and encountered an error bridging Chameleon into swift.

The Swift compiler did not recognize any of the UIKit classes in the ChamelonStatusBar class or the NSArray+Chameleon category.

I was able to fix it by importing UIKit in the headers for those classes.
After doing that I had no problems building.